requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.3
